requirements.txt
actions==1.0.3
allure-pytest==2.12.0
allure-python-commons==2.12.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
basepage==1.7.4
bunch==1.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
docutils==0.19
exceptiongroup==1.0.4
Faker==16.6.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
MarkupSafe==2.1.2
outcome==1.2.0
packaging==21.3
pages==0.3
piglet-templates==1.3.0
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.5
wsproto==1.2.0
zipp==3.12.0